refactor: rename demographic.py
