+++ backend/tests/engine/core/test_indicators.py
@@ -27,13 +27,13 @@ def test_calculate_gdp():
     assert False
 
 
-@mark.xfail(reason="Demographic's Income feature not implemented yet.")
+@mark.xfail(reason="Demographic's income feature not implemented yet.")
 def test_calculate_income_per_capita(model):
     # TODO: do whenever templates for demographic's distribution has been done.
     assert False
 
 
-@mark.xfail(reason="Demographic's Income feature  not implemented yet.")
+@mark.xfail(reason="Demographic's income feature not implemented yet.")
 def test_calculate_median_income(model):
     # TODO: do whenever templates for demographic's distribution has been done.
     assert False
